Checked test requirements

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int answer = toBeDivided / 5;
+        float answer = toBeDivided/ 5;
         System.out.println(answer);
     }
 }
